+++ mne_bids/tests/test_path.py
@@ -23,7 +23,8 @@ from mne_bids import (get_kinds, get_entity_vals, print_dir_tree,
                       make_bids_folders, make_bids_basename,
                       write_raw_bids)
 from mne_bids.path import (_parse_ext, get_entities_from_fname,
-                           _find_best_candidates, _find_matching_sidecar)
+                           _find_best_candidates, _find_matching_sidecar,
+                           _filter_fnames, get_matched_basenames)
 
 subject_id = '01'
 session_id = '01'
@@ -335,3 +336,50 @@ def test_bids_path(return_bids_test_dir):
                                    task='03', kind='ieeg',
                                    extension='.edf')
     assert repr(bids_path) == 'BIDSPath(sub-01_ses-02_task-03_ieeg.edf)'
+
+
+@pytest.mark.parametrize(
+    'entities, expected_n_matches',
+    [
+        (dict(), 9),
+        (dict(subject='01'), 2),
+        (dict(task='audio'), 2),
+        (dict(processing='sss'), 1),
+        (dict(kind='meg'), 4),
+        (dict(acquisition='t1w'), 1),
+        (dict(task='test', processing='ica', kind='eeg'), 2),
+        (dict(subject='5', task='test', processing='ica', kind='eeg'), 1)
+    ])
+def test_filter_fnames(entities, expected_n_matches):
+    """Test filtering filenames based on BIDS entities works."""
+
+    fnames = ('sub-01_task-audio_meg.fif',
+              'sub-01_ses-05_task-audio_meg.fif',
+              'sub-02_task-visual_eeg.vhdr',
+              'sub-Foo_ses-bar_meg.fif',
+              'sub-Bar_task-invasive_run-1_ieeg.fif',
+              'sub-3_task-fun_proc-sss_meg.fif',
+              'sub-4_task-pain_acq-t1w_mri.nii.gz',
+              'sub-5_task-test_proc-ica_eeg.vhdr',
+              'sub-6_task-test_proc-ica_eeg.vhdr')
+
+    output = _filter_fnames(fnames, **entities)
+    assert len(output) == expected_n_matches
+
+
+def test_get_matched_basenames(return_bids_test_dir):
+    """Test retrieval of matching basenames."""
+    bids_root = return_bids_test_dir
+
+    paths = get_matched_basenames(bids_root=bids_root)
+    assert len(paths) == 2
+    assert paths[0].basename == 'sub-01_ses-01_task-testing_run-01_meg'
+    assert paths[1].basename == 'sub-01_ses-01_task-testing_run-02_meg'
+    assert all([p.prefix == bids_root for p in paths])
+
+    paths = get_matched_basenames(bids_root=bids_root, run='01')
+    assert len(paths) == 1
+    assert paths[0].basename == 'sub-01_ses-01_task-testing_run-01_meg'
+
+    paths = get_matched_basenames(bids_root=bids_root, subject='unknown')
+    assert len(paths) == 0